requirements.txt
ibm-watson==10.0.0
ibm-cloud-sdk-core==3.24.2
gradio==5.38.2
Pillow==11.3.0
requests==2.32.4
numpy==2.3.2 